Week4/feature/coupon 도메인 추가
* fix: 쿠폰 설계 추가

* feat: 쿠폰 목록 조회 기능 구현

* feat: 쿠폰 목록 조회 테스트 구현

* feat: 주문 쿠폰 적용

+++ apps/commerce-api/src/main/java/com/loopers/application/point/PointChargeInfo.java
@@ -3,11 +3,13 @@ package com.loopers.application.point;
 import com.loopers.domain.point.PointModel;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 public class PointChargeInfo {
-    private int point;
+    private BigDecimal point;
 
-    public PointChargeInfo(int point) {
+    public PointChargeInfo(BigDecimal point) {
         this.point = point;
     }
